+++ fvm/src/kernel/mod.rs
@@ -14,7 +14,8 @@ use fvm_shared::sector::{
     AggregateSealVerifyProofAndInfos, RegisteredSealProof, ReplicaUpdateInfo, SealVerifyInfo,
     WindowPoStVerifyInfo,
 };
-use fvm_shared::version::NetworkVersion;
+use fvm_shared::sys::out::network::NetworkContext;
+use fvm_shared::sys::out::vm::MessageContext;
 use fvm_shared::{ActorID, MethodNum};
 
 mod hash;
@@ -34,9 +35,10 @@ use crate::gas::{Gas, PriceList};
 use crate::machine::limiter::ExecMemory;
 use crate::machine::Machine;
 
-pub enum SendResult {
-    Return(BlockId, BlockStat),
-    Abort(ExitCode),
+pub struct SendResult {
+    pub block_id: BlockId,
+    pub block_stat: BlockStat,
+    pub exit_code: ExitCode,
 }
 
 /// The "kernel" implements the FVM interface as presented to the actors. It:
@@ -95,17 +97,8 @@ pub trait Kernel:
 
 /// Network-related operations.
 pub trait NetworkOps {
-    /// The current network epoch (constant).
-    fn network_epoch(&self) -> ChainEpoch;
-
-    /// The current network version (constant).
-    fn network_version(&self) -> NetworkVersion;
-
-    /// The current base-fee (constant).
-    fn network_base_fee(&self) -> &TokenAmount;
-
-    /// The current tipset timestamp (seconds since the unix epoch).
-    fn tipset_timestamp(&self) -> u64;
+    /// Network information (epoch, version, etc.).
+    fn network_context(&self) -> Result<NetworkContext>;
 
     /// The CID of the tipset at the specified epoch.
     fn tipset_cid(&self, epoch: ChainEpoch) -> Result<Cid>;
@@ -113,26 +106,8 @@ pub trait NetworkOps {
 
 /// Accessors to query attributes of the incoming message.
 pub trait MessageOps {
-    /// The calling actor (constant).
-    fn msg_caller(&self) -> ActorID;
-
-    /// The origin actor
-    fn msg_origin(&self) -> ActorID;
-
-    /// The receiving actor (this actor) (constant).
-    fn msg_receiver(&self) -> ActorID;
-
-    /// The method number used to invoke this actor (constant).
-    fn msg_method_number(&self) -> MethodNum;
-
-    /// The value received from the caller (constant).
-    fn msg_value_received(&self) -> TokenAmount;
-
-    /// The current message gas premium
-    fn msg_gas_premium(&self) -> TokenAmount;
-
-    /// The current message gas limit
-    fn msg_gas_limit(&self) -> u64;
+    /// Message information.
+    fn msg_context(&self) -> Result<MessageContext>;
 }
 
 /// The IPLD subset of the kernel.
@@ -159,12 +134,12 @@ pub trait IpldBlockOps {
     /// Read data from a block.
     ///
     /// This method will fail if the block handle is invalid.
-    fn block_read(&mut self, id: BlockId, offset: u32, buf: &mut [u8]) -> Result<i32>;
+    fn block_read(&self, id: BlockId, offset: u32, buf: &mut [u8]) -> Result<i32>;
 
     /// Returns the blocks codec & size.
     ///
     /// This method will fail if the block handle is invalid.
-    fn block_stat(&mut self, id: BlockId) -> Result<BlockStat>;
+    fn block_stat(&self, id: BlockId) -> Result<BlockStat>;
 }
 
 /// Actor state access and manipulation.
@@ -205,7 +180,7 @@ pub trait ActorOps {
     /// the actor even in the event of a chain re-org (whereas an ID-address might refer to a
     /// different actor after messages are re-ordered).
     /// Always an ActorExec address.
-    fn new_actor_address(&mut self) -> Result<Address>;
+    fn next_actor_address(&self) -> Result<Address>;
 
     /// Creates an actor with given `code_cid`, `actor_id`, `predictable_address` (if specified),
     /// and an empty state.
@@ -264,7 +239,7 @@ pub trait GasOps {
 
     /// ChargeGas charges specified amount of `gas` for execution.
     /// `name` provides information about gas charging point.
-    fn charge_gas(&mut self, name: &str, compute: Gas) -> Result<()>;
+    fn charge_gas(&self, name: &str, compute: Gas) -> Result<()>;
 
     /// Returns the currently active gas price list.
     fn price_list(&self) -> &PriceList;
@@ -274,7 +249,7 @@ pub trait GasOps {
 pub trait CryptoOps {
     /// Verifies that a signature is valid for an address and plaintext.
     fn verify_signature(
-        &mut self,
+        &self,
         sig_type: SignatureType,
         signature: &[u8],
         signer: &Address,
@@ -283,7 +258,7 @@ pub trait CryptoOps {
 
     /// Given a message hash and its signature, recovers the public key of the signer.
     fn recover_secp_public_key(
-        &mut self,
+        &self,
         hash: &[u8; SECP_SIG_MESSAGE_HASH_SIZE],
         signature: &[u8; SECP_SIG_LEN],
     ) -> Result<[u8; SECP_PUB_LEN]>;
@@ -292,20 +267,20 @@ pub trait CryptoOps {
     /// `digest_out`, returning the size of the digest written to `digest_out`. If `digest_out` is
     /// to small to fit the entire digest, it will be truncated. If too large, the leftover space
     /// will not be overwritten.
-    fn hash(&mut self, code: u64, data: &[u8]) -> Result<MultihashGeneric<64>>;
+    fn hash(&self, code: u64, data: &[u8]) -> Result<MultihashGeneric<64>>;
 
     /// Computes an unsealed sector CID (CommD) from its constituent piece CIDs (CommPs) and sizes.
     fn compute_unsealed_sector_cid(
-        &mut self,
+        &self,
         proof_type: RegisteredSealProof,
         pieces: &[PieceInfo],
     ) -> Result<Cid>;
 
     /// Verifies a sector seal proof.
-    fn verify_seal(&mut self, vi: &SealVerifyInfo) -> Result<bool>;
+    fn verify_seal(&self, vi: &SealVerifyInfo) -> Result<bool>;
 
     /// Verifies a window proof of spacetime.
-    fn verify_post(&mut self, verify_info: &WindowPoStVerifyInfo) -> Result<bool>;
+    fn verify_post(&self, verify_info: &WindowPoStVerifyInfo) -> Result<bool>;
 
     /// Verifies that two block headers provide proof of a consensus fault:
     /// - both headers mined by the same actor
@@ -318,7 +293,7 @@ pub trait CryptoOps {
     /// blocks in the parent of h2 (i.e. h2's grandparent).
     /// Returns nil and an error if the headers don't prove a fault.
     fn verify_consensus_fault(
-        &mut self,
+        &self,
         h1: &[u8],
         h2: &[u8],
         extra: &[u8],
@@ -329,17 +304,14 @@ pub trait CryptoOps {
     ///
     /// Gas: This syscall intentionally _does not_ charge any gas (as said gas would be charged to
     /// cron). Instead, gas is pre-paid by the storage provider on pre-commit.
-    fn batch_verify_seals(&mut self, vis: &[SealVerifyInfo]) -> Result<Vec<bool>>;
+    fn batch_verify_seals(&self, vis: &[SealVerifyInfo]) -> Result<Vec<bool>>;
 
     /// Verify aggregate seals verifies an aggregated batch of prove-commits.
-    fn verify_aggregate_seals(
-        &mut self,
-        aggregate: &AggregateSealVerifyProofAndInfos,
-    ) -> Result<bool>;
+    fn verify_aggregate_seals(&self, aggregate: &AggregateSealVerifyProofAndInfos) -> Result<bool>;
 
     /// Verify replica update verifies a snap deal: an upgrade from a CC sector to a sector with
     /// deals.
-    fn verify_replica_update(&mut self, replica: &ReplicaUpdateInfo) -> Result<bool>;
+    fn verify_replica_update(&self, replica: &ReplicaUpdateInfo) -> Result<bool>;
 }
 
 /// Randomness queries.
@@ -348,7 +320,7 @@ pub trait RandomnessOps {
     /// ticket chain from a given epoch and incorporating requisite entropy.
     /// This randomness is fork dependant but also biasable because of this.
     fn get_randomness_from_tickets(
-        &mut self,
+        &self,
         personalization: i64,
         rand_epoch: ChainEpoch,
         entropy: &[u8],
@@ -358,7 +330,7 @@ pub trait RandomnessOps {
     /// beacon from a given epoch and incorporating requisite entropy.
     /// This randomness is not tied to any fork of the chain, and is unbiasable.
     fn get_randomness_from_beacon(
-        &mut self,
+        &self,
         personalization: i64,
         rand_epoch: ChainEpoch,
         entropy: &[u8],