+++ fvm/src/syscalls/mod.rs
@@ -157,18 +157,15 @@ use self::error::Abort;
 pub fn bind_syscalls(
     linker: &mut Linker<InvocationData<impl Kernel + 'static>>,
 ) -> anyhow::Result<()> {
-    linker.bind("vm", "abort", vm::abort)?;
+    linker.bind("vm", "exit", vm::exit)?;
     linker.bind("vm", "message_context", vm::message_context)?;
 
-    linker.bind("network", "base_fee", network::base_fee)?;
     linker.bind(
         "network",
         "total_fil_circ_supply",
         network::total_fil_circ_supply,
     )?;
     linker.bind("network", "context", network::context)?;
-
-    linker.bind("network", "tipset_timestamp", network::tipset_timestamp)?;
     linker.bind("network", "tipset_cid", network::tipset_cid)?;
 
     linker.bind("ipld", "block_open", ipld::block_open)?;
@@ -185,7 +182,7 @@ pub fn bind_syscalls(
     linker.bind("actor", "resolve_address", actor::resolve_address)?;
     linker.bind("actor", "lookup_address", actor::lookup_address)?;
     linker.bind("actor", "get_actor_code_cid", actor::get_actor_code_cid)?;
-    linker.bind("actor", "new_actor_address", actor::new_actor_address)?;
+    linker.bind("actor", "next_actor_address", actor::next_actor_address)?;
     linker.bind("actor", "create_actor", actor::create_actor)?;
     linker.bind(
         "actor",