+++ testing/integration/tests/address_test.rs
@@ -11,9 +11,8 @@ use fvm_shared::state::StateTreeVersion;
 use fvm_shared::version::NetworkVersion;
 use num_traits::Zero;
 
-// XXX RE-ENABLE THIS AFTER MERGE WITH BUILTIN ACTORS
-// #[test]
-fn _basic_address_tests() {
+#[test]
+fn basic_address_tests() {
     // Instantiate tester
     let mut tester = new_tester(
         NetworkVersion::V18,